requirements.txt
# 核心框架
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0

# Web 框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0

# 資料庫 (SQLite)
sqlalchemy>=2.0.0

# 環境變數管理
python-dotenv>=1.0.0

# 資料處理
pydantic>=2.5.0

# 工具庫
requests>=2.31.0
python-multipart>=0.0.6

# 開發工具
pytest>=7.4.0
black>=23.0.0

# 日誌
loguru>=0.7.0
